# PERF: Calculate mask in interpolate only once