--- a/pandas/core/generic.py
+++ b/pandas/core/generic.py
@@ -11643,7 +11643,7 @@ class NDFrame(PandasObject, indexing.IndexingMixin):
         -------
         idx_first_valid : type of index
         """
-        idxpos = find_valid_index(self._values, how=how)
+        idxpos = find_valid_index(self._values, how=how, is_valid=isna(self._values))
         if idxpos is None:
             return None
         return self.index[idxpos]