--- a/pandas/core/missing.py
+++ b/pandas/core/missing.py
@@ -170,7 +170,7 @@ def clean_interp_method(method: str, index: Index, **kwargs) -> str:
     return method
 
 
-def find_valid_index(values, *, how: str) -> int | None:
+def find_valid_index(values, *, how: str, is_valid: np.ndarray) -> int | None:
     """
     Retrieves the index of the first valid value.
 
@@ -179,6 +179,8 @@ def find_valid_index(values, *, how: str) -> int | None:
     values : ndarray or ExtensionArray
     how : {'first', 'last'}
         Use this parameter to change between the first or last valid index.
+    is_valid: np.ndarray
+        Mask to find na_values.
 
     Returns
     -------
@@ -188,8 +190,6 @@ def find_valid_index(values, *, how: str) -> int | None:
 
     if len(values) == 0:  # early stop
         return None
-
-    is_valid = ~isna(values)
 
     if values.ndim == 2:
         is_valid = is_valid.any(axis=1)  # reduce axis 1
@@ -400,12 +400,12 @@ def _interpolate_1d(
     # These are sets of index pointers to invalid values... i.e. {0, 1, etc...
     all_nans = set(np.flatnonzero(invalid))
 
-    first_valid_index = find_valid_index(yvalues, how="first")
+    first_valid_index = find_valid_index(yvalues, how="first", is_valid=valid)
     if first_valid_index is None:  # no nan found in start
         first_valid_index = 0
     start_nans = set(range(first_valid_index))
 
-    last_valid_index = find_valid_index(yvalues, how="last")
+    last_valid_index = find_valid_index(yvalues, how="last", is_valid=valid)
     if last_valid_index is None:  # no nan found in end
         last_valid_index = len(yvalues)
     end_nans = set(range(1 + last_valid_index, len(valid)))
@@ -738,12 +738,13 @@ def _interpolate_with_limit_area(
     """
 
     invalid = isna(values)
+    is_valid = ~invalid
 
     if not invalid.all():
-        first = find_valid_index(values, how="first")
+        first = find_valid_index(values, how="first", is_valid=is_valid)
         if first is None:
             first = 0
-        last = find_valid_index(values, how="last")
+        last = find_valid_index(values, how="last", is_valid=is_valid)
         if last is None:
             last = len(values)